WTF
git-svn-id: https://repo.acin.tuwien.ac.at/v4r/stable/strands_v4r@9167 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ v4r/ORGTEvaluator/main_window.h
@@ -146,7 +146,7 @@ public:
    {
        for (size_t i = 0; i < single_scenes_.size (); i++)
        {
-           if( static_cast<int>(i) == selected_scene_)
+           if( i == selected_scene_)
            {
                std::stringstream cloud_name;
                cloud_name << "view_" << i;
